requirements.txt
openai
fastapi
pydantic
uvicorn
langchain-openai
langchain-community
pypdf
cryptography
langchain-core
langgraph
beautifulsoup4
IPython
langchain-chroma
langchain-weaviate
weaviate
weaviate-client
simsimd
